refactor: 즐겨찾기 전용 에러 추가
-인증토큰의 회원정보가 없는 경우 에러발생 (401에러)
