+++ src/main/java/nextstep/member/exception/FavoriteRestApiException.java
@@ -1,0 +1,10 @@
+package nextstep.member.exception;
+
+public class FavoriteRestApiException extends RuntimeException{
+    public FavoriteRestApiException() {
+    }
+
+    public FavoriteRestApiException(String message) {
+        super(message);
+    }
+}
